Update Spring 3.3 and JUnit5 migration

+++ src/test/java/gov/nist/oar/distrib/cachemgr/simple/SimpleCacheManagerTest.java
@@ -30,10 +30,9 @@ import org.junit.jupiter.api.io.TempDir;
 
 import gov.nist.oar.distrib.LongTermStorage;
 import gov.nist.oar.distrib.cachemgr.CacheManagementException;
-import gov.nist.oar.distrib.cachemgr.CacheManager;
 import gov.nist.oar.distrib.cachemgr.CacheObject;
 import gov.nist.oar.distrib.cachemgr.CacheVolume;
-import gov.nist.oar.distrib.cachemgr.InventoryException;
+import gov.nist.oar.distrib.cachemgr.DeletionPlanner;
 import gov.nist.oar.distrib.cachemgr.StorageInventoryDB;
 import gov.nist.oar.distrib.cachemgr.inventory.DefaultDeletionPlanner;
 import gov.nist.oar.distrib.cachemgr.inventory.OldSelectionStrategy;
@@ -42,10 +41,10 @@ import gov.nist.oar.distrib.cachemgr.restore.FileCopyRestorer;
 import gov.nist.oar.distrib.cachemgr.storage.FilesystemCacheVolume;
 import gov.nist.oar.distrib.storage.FilesystemLongTermStorage;
 
-class SimpleCacheManagerTest {
+public class SimpleCacheManagerTest {
 
     @TempDir
-    File tempDir;  // JUnit 5's @TempDir for managing temporary files
+    public File tempDir;  // JUnit 5's @TempDir for managing temporary files
 
     File dbf = null;
     File repodir = null;
@@ -53,26 +52,24 @@ class SimpleCacheManagerTest {
     List<CacheVolume> cvlist = null;
     SimpleCache cache = null;
 
-    String createDB() throws IOException, InventoryException {
-        File tf = new File(tempDir, "testdb.sqlite");
-        String out = tf.getAbsolutePath();
-        SQLiteStorageInventoryDB.initializeDB(out);
-        return out;
+    String createDB() throws IOException, CacheManagementException {
+        dbf = new File(tempDir, "testdb.sqlite");
+        SQLiteStorageInventoryDB.initializeDB(dbf.getAbsolutePath());
+        return dbf.getAbsolutePath();
     }
 
     @AfterEach
-    void tearDown() {
-        if (dbf != null) {
+    public void tearDown() {
+        if (dbf != null && dbf.exists()) {
             dbf.delete();
         }
     }
 
     @BeforeEach
-    void setUp() throws IOException, InventoryException {
-        dbf = new File(createDB());
-        assertTrue(dbf.exists());
+    public void setUp() throws IOException, CacheManagementException {
+        createDB();
 
-        sidb = new SQLiteStorageInventoryDB(dbf.getPath());
+        sidb = new SQLiteStorageInventoryDB(dbf.getAbsolutePath());
         sidb.registerAlgorithm("md5");
         sidb.registerAlgorithm("sha256");
 
@@ -80,55 +77,52 @@ class SimpleCacheManagerTest {
         for (int i = 0; i < 3; i++) {
             String name = "Cache" + i;
             File cd = new File(tempDir, name);
-            CacheVolume cv = new FilesystemCacheVolume(cd.toString(), name);
+            cd.mkdir();
+            CacheVolume cv = new FilesystemCacheVolume(cd.getAbsolutePath(), name);
             sidb.registerVolume(name, 120000, null);
             cvlist.add(cv);
         }
 
         OldSelectionStrategy ss = new OldSelectionStrategy(120000, 120000, 0, 0);
-        cache = new SimpleCache("Simple", sidb, cvlist, new DefaultDeletionPlanner(sidb, cvlist, ss));
+        DeletionPlanner dp = new DefaultDeletionPlanner(sidb, cvlist, ss);
+        cache = new SimpleCache("Simple", sidb, cvlist, dp);
     }
 
-    File createRepo() throws IOException {
+    public File createRepo() throws IOException {
         repodir = new File(tempDir, "repo");
         repodir.mkdir();
-
         String[] zips = {
             "mds1491.mbag0_2-0.zip", "mds1491.1_1_0.mbag0_4-1.zip",
             "67C783D4BA814C8EE05324570681708A1899.mbag0_3-0.zip",
             "67C783D4BA814C8EE05324570681708A1899.mbag0_3-1.zip"
         };
-
         for (String file : zips) {
-            Files.copy(ZipRepoRestorerTest.class.getResourceAsStream("/" + file),
-                       new File(repodir, file).toPath());
-            Files.copy(ZipRepoRestorerTest.class.getResourceAsStream("/" + file + ".sha256"),
-                       new File(repodir, file + ".sha256").toPath());
+            Files.copy(getClass().getResourceAsStream("/" + file),
+                new File(repodir, file).toPath());
+            Files.copy(getClass().getResourceAsStream("/" + file + ".sha256"),
+                new File(repodir, file + ".sha256").toPath());
         }
-
-        Files.copy(ZipRepoRestorerTest.class.getResourceAsStream("/67C783D4BA814C8EE05324570681708A1899.mbag0_3-0.zip"),
-                   new File(repodir, "goober.zip").toPath());
-        Files.copy(ZipRepoRestorerTest.class.getResourceAsStream("/67C783D4BA814C8EE05324570681708A1899.mbag0_3-0.zip.sha256"),
-                   new File(repodir, "goober.zip.sha256").toPath());
+        Files.copy(getClass().getResourceAsStream("/67C783D4BA814C8EE05324570681708A1899.mbag0_3-0.zip"),
+            new File(repodir, "goober.zip").toPath());
+        Files.copy(getClass().getResourceAsStream("/67C783D4BA814C8EE05324570681708A1899.mbag0_3-0.zip.sha256"),
+            new File(repodir, "goober.zip.sha256").toPath());
 
         return repodir;
     }
 
     @Test
-    void testCtor() throws IOException, CacheManagementException {
+    public void testCtor() throws IOException, CacheManagementException {
         LongTermStorage lts = new FilesystemLongTermStorage(createRepo().toString());
         SimpleCacheManager scm = new SimpleCacheManager(cache, new FileCopyRestorer(lts));
-
         for (Long space : sidb.getAvailableSpace().values()) {
             assertEquals(120000L, space.longValue());
         }
-
         assertFalse(scm.isCached("mds1491.mbag0_2-0.zip"));
         scm.uncache("mds1491.mbag0_2-0.zip");
     }
 
     @Test
-    void testCache() throws IOException, CacheManagementException {
+    public void testCache() throws IOException, CacheManagementException {
         LongTermStorage lts = new FilesystemLongTermStorage(createRepo().toString());
         SimpleCacheManager scm = new SimpleCacheManager(cache, new FileCopyRestorer(lts));
 
@@ -148,7 +142,7 @@ class SimpleCacheManagerTest {
     }
 
     @Test
-    void testFindObject() throws IOException, CacheManagementException {
+    public void testFindObject() throws IOException, CacheManagementException {
         LongTermStorage lts = new FilesystemLongTermStorage(createRepo().toString());
         SimpleCacheManager scm = new SimpleCacheManager(cache, new FileCopyRestorer(lts));
 
@@ -158,16 +152,16 @@ class SimpleCacheManagerTest {
         assertTrue(scm.isCached("mds1491.mbag0_2-0.zip"));
 
         CacheObject co = scm.findObject("mds1491.mbag0_2-0.zip");
-        assertTrue(co.volname.startsWith("Cache"), "Unexpected cache name: " + co.volname);
+        assertTrue(co.volname.startsWith("Cache"));
         assertEquals("mds1491.mbag0_2-0.zip", co.name);
         assertNotNull(co.volume);
 
         assertEquals(9841, co.getSize());
         assertEquals("3de9d9e32831be693e341306db79e636ebd61b6a78f9482e8e3038a6e8eba569",
-                     co.getMetadatumString("checksum", "not-set"));
+                co.getMetadatumString("checksum", "not-set"));
 
         File cf = new File(new File(tempDir, co.volname), "mds1491.mbag0_2-0.zip");
-        assertTrue(cf.exists(), "Can't find cache file");
+        assertTrue(cf.exists());
 
         long accesstime = co.getMetadatumLong("since", -1L);
         assertTrue(accesstime > 0);
@@ -177,7 +171,7 @@ class SimpleCacheManagerTest {
     }
 
     @Test
-    void testGetObject() throws IOException, CacheManagementException {
+    public void testGetObject() throws IOException, CacheManagementException {
         LongTermStorage lts = new FilesystemLongTermStorage(createRepo().toString());
         SimpleCacheManager scm = new SimpleCacheManager(cache, new FileCopyRestorer(lts));
 
@@ -186,11 +180,11 @@ class SimpleCacheManagerTest {
         assertTrue(scm.isCached("mds1491.mbag0_2-0.zip"));
 
         File cf = new File(new File(tempDir, co.volname), "mds1491.mbag0_2-0.zip");
-        assertTrue(cf.exists(), "Can't find cache file");
+        assertTrue(cf.exists());
     }
 
     @Test
-    void testFill1() throws IOException, CacheManagementException {
+    public void testFill1() throws IOException, CacheManagementException {
         LongTermStorage lts = new FilesystemLongTermStorage(createRepo().toString());
         SimpleCacheManager scm = new SimpleCacheManager(cache, new FileCopyRestorer(lts));
 
@@ -205,10 +199,10 @@ class SimpleCacheManagerTest {
         assertTrue(scm.isCached("goober.zip"));
 
         File cf = new File(new File(tempDir, co.volname), "goober.zip");
-        assertTrue(cf.exists(), "Can't find cache file");
+        assertTrue(cf.exists());
 
         // something had to have gotten deleted
-        int i = 0;
+        int i;
         for (i = 0; i < zips.length; i++) {
             if (!scm.isCached(zips[i])) break;
         }
@@ -216,7 +210,7 @@ class SimpleCacheManagerTest {
     }
 
     @Test
-    void testFill2() throws IOException, CacheManagementException {
+    public void testFill2() throws IOException, CacheManagementException {
         LongTermStorage lts = new FilesystemLongTermStorage(createRepo().toString());
         SimpleCacheManager scm = new SimpleCacheManager(cache, new FileCopyRestorer(lts));
 
@@ -229,7 +223,7 @@ class SimpleCacheManagerTest {
         testFill(scm, zips);
     }
 
-    void testFill(CacheManager scm, String[] zips) throws IOException, CacheManagementException {
+    public void testFill(SimpleCacheManager scm, String[] zips) throws IOException, CacheManagementException {
         assertFalse(scm.isCached("mds1491.mbag0_2-0.zip"));
 
         for (String file : zips) {
@@ -242,7 +236,7 @@ class SimpleCacheManagerTest {
             assertTrue(scm.isCached(co.id));
 
             File cf = new File(new File(tempDir, co.volname), co.name);
-            assertTrue(cf.exists(), "Can't find cache file");
+            assertTrue(cf.exists());
         }
 
         for (Long space : sidb.getAvailableSpace().values()) {
